pyproject.toml
[tool.poetry]
name = "ardio"
version = "1.1.0"
description = "Journal article to audio book"
authors = ["Bell Eapen <github@gulfdoctor.net>"]

readme = 'README.md'  # Markdown files are supported

repository = "https://github.com/dermatologist/ardio"
homepage = "https://github.com/dermatologist/ardio"

classifiers = [
    "Intended Audience :: Science/Research",
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.7",
    "Topic :: Scientific/Engineering :: Information Analysis"
]

[tool.poetry.dependencies]
python = "^3.7"
"pdfminer.six" = "^20201018"
gTTS = "^1.2.2"
gTTS-token = "^1.1.4"

[tool.poetry.dev-dependencies]
autopep8 = "^1.5.4"

[tool.poetry.scripts]
ardio = "ardio.ardio:main_routine"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
